Seperating categories and pill into new files
